+++ Alchemy/media/audio/metadata.hxx
@@ -7,19 +7,19 @@
 namespace Alchemist::Media::Audio {
 	class ALCHEMY_PUBLIC Metadata: public Media::Metadata {
 		public:
-			Metadata(const unsigned short&, const unsigned short&);
-			Metadata(unsigned short&&, unsigned short&&) noexcept;
+			Metadata(const int&, const int&);
+			Metadata(int&&, int&&) noexcept;
 			Metadata(const Metadata&)					= default;
 			Metadata(Metadata&&) noexcept				= default;
 			Metadata& operator=(const Metadata&)		= default;
 			Metadata& operator=(Metadata&&) noexcept	= default;
 			~Metadata() noexcept						= default;
 
-			const unsigned short& 	GetChannels() const noexcept;
-			const unsigned short& 	GetSampleRate() const noexcept;
+			const int& 	GetChannels() const noexcept;
+			const int& 	GetSampleRate() const noexcept;
 			std::string 			GetString() const override;
 
 		private:
-			unsigned short m_channels, m_sample_rate;
+			int m_channels, m_sample_rate;
 	};
 }
