+++ Alchemy/media/file.hxx
@@ -22,12 +22,12 @@ namespace Alchemist::Media {
 
 			const std::filesystem::path&				GetFileName() const noexcept;
 			const std::vector<std::shared_ptr<Stream>>&	GetStreams() const noexcept;
-			const unsigned long long&					GetSize() const noexcept;
+			const uint64_t&								GetSize() const noexcept;
 
 		private:
 			std::filesystem::path m_filename;
 			std::vector<std::shared_ptr<Stream>> m_streams;
-			unsigned long long m_size_bytes;
+			uint64_t m_size_bytes;
 
 			void InitStreams();
 			std::shared_ptr<Stream> 		ParseAudioInfo(const Json::Value&);
