+++ Alchemy/media/stream.hxx
@@ -10,15 +10,15 @@
 namespace Alchemist::Media {
 	class ALCHEMY_PUBLIC Stream {
 		public:
-			Stream(const unsigned short&, const Type&) noexcept;
+			Stream(const int&, const Type&) noexcept;
 			Stream(const Stream&)					= default;
 			Stream(Stream&&) noexcept				= default;
 			Stream& operator=(const Stream&)		= default;
 			Stream& operator=(Stream&&) noexcept	= default;
 			virtual ~Stream() noexcept				= 0;
 
-			const unsigned short&						GetIndex() const noexcept;
-			void										SetIndex(const unsigned short&) noexcept;
+			const int&						GetIndex() const noexcept;
+			void										SetIndex(const int&) noexcept;
 			const Type&									GetType() const noexcept;
 			const std::optional<std::string>& 			GetLanguage() const noexcept;
 			void 										SetLanguage(const std::string&);
@@ -33,7 +33,7 @@ namespace Alchemist::Media {
 			const std::shared_ptr<Metadata>&			GetMetadata() const noexcept;
 			
 		protected:
-			unsigned short m_index;
+			int m_index;
 			std::optional<std::string> m_language, m_title;
 			bool m_default, m_forced;
 			std::shared_ptr<Codec> m_codec;
