+++ Alchemy/media/type.hxx
@@ -3,7 +3,7 @@
 #include <Alchemy/visibility.h>
 
 namespace Alchemist::Media {
-	enum class ALCHEMY_PUBLIC Type: unsigned short {
+	enum class ALCHEMY_PUBLIC Type: int {
 		Audio		= 1,
 		Video		= 2,
 		Subtitle	= 3
