+++ Alchemy/media/video/stream.cxx
@@ -2,7 +2,7 @@
 
 using namespace Alchemist::Media::Video;
 
-Stream::Stream(const unsigned short& index) noexcept:Media::Stream(index, Type::Video) {}
+Stream::Stream(const int& index) noexcept:Media::Stream(index, Type::Video) {}
 
 void Stream::SetMetadata(const Metadata& meta) { m_metadata = std::make_shared<Metadata>(meta); }
